docs: Add lesson learned for egui slider reset
Adds a 'Lesson Learned' entry to BUGS.md detailing the resolution of the slider reset bug. This documents the correct, idiomatic way to use `nih_plug_egui::widgets::ParamSlider` to ensure UI controls are properly integrated with the `nih-plug` parameter system.

This also marks the corresponding documentation task as complete in TODO.md.